pyproject.toml
[tool.poetry]
name = "ut-telegram-bot"
version = "4.6.1"
description = ""
authors = ["Habib Rohman <habibrohman@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.8"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
beautifulsoup4 = "^4.9.1"
lxml = "^4.5.2"
dacite = "^1.5.1"
cachetools = "^4.1.1"
fuzzywuzzy = "^0.18.0"
Flask = "^1.1.2"
flask_sqlalchemy = "^2.4.4"
flask_migrate = "^2.5.3"
gunicorn = "^20.0.4"
flask_script = "^2.0.6"
ujson = "^3.1.0"
PyYAML = "^5.3.1"
python-Levenshtein = "^0.12.0"
pydantic = "^1.6.1"
bleach = "^3.1.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.30.0"
flake8 = "^3.8.3"
pre-commit = "^2.7.1"
rope = "^0.17.0"
tox = "^3.20.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tox]
envlist = "py37"
